Update 03. Area Of Rectangle.java

+++ Problems-&-Solutions/Formula Related Problems/03. Area Of Rectangle.java
@@ -9,7 +9,7 @@ class Main {
         int l = scan.nextInt();
         System.out.print("Enter breadth of the rectangle :");
         int b = scan.nextInt();
-        double area = l*b;
+        double area = (double)l*b;
         System.out.println("Area of a rectangle : "+area);
     }
 }
